requirements.txt
tensorflow
pandas
numpy 
Pillow
opencv-python
scikit-learn
matplotlib
seaborn
tqdm
-e .
